apps/api/requirements.txt
fastapi==0.115.6
mangum==0.19.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
anthropic==0.44.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
tenacity==9.0.0
requests==2.32.3
openpyxl==3.1.5
pandas==2.2.3
pytest==8.3.4
pytest-asyncio==0.25.0
